+++ CHANGES.md
@@ -2,9 +2,16 @@
 
 ## Unreleased
 
+### titiler.core
+* fix template loading for starlette >= 0.28 by using `jinja2.Environment` argument
+
+### titiler.extensions
+* fix template loading for starlette >= 0.28 by using `jinja2.Environment` argument
+
 ### titiler.application
 
 * add simple *auth* (optional) based on `global_access_token` string, set with `TITILER_API_GLOBAL_ACCESS_TOKEN` environment variable (author @DeflateAwning, https://github.com/developmentseed/titiler/pull/735)
+* fix template loading for starlette >= 0.28 by using `jinja2.Environment` argument
 
 ## 0.15.6 (2023-11-16)